27th commit

+++ db/migrate/20230904025333_remove_post_id_from_response_likes.rb
@@ -1,5 +1,7 @@
 class RemovePostIdFromResponseLikes < ActiveRecord::Migration[6.0]
   def change
-    remove_column :response_likes, :post_id, :integer
-  end
+    if column_exists?(:response_likes, :post_id)
+      remove_column :response_likes, :post_id, :integer
+    end
+  end  
 end
